# back/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TECLAR — API",
    "version": "1.0",
    "description": "Gerada do mesmo contrato que o CONTRATO-API.md (js/nucleo/api.contrato.ts), por gerar-contrato.mjs. Rotas marcadas \"(ainda não implementada)\" ainda não existem no back.\n\nPara testar uma rota com cadeado: faça o POST /api/auth/login, copie o \"token\" da resposta e cole em **Authorize** (sem a palavra Bearer)."
  },
...
  "paths": {
...
    "/api/turmas/{id}/atribuicoes": {
      "get": {
        "tags": [
          "Turmas (professor)"
        ],
        "operationId": "turmas.atribuicoes",
        "summary": "atribuicoes: AtribuicaoProfessor[] — o que a turma recebeu, na ordem de atribuição",
        "description": "**Chamada no front:** `api.turmas.atribuicoes`\n\n**Resposta:** 200 AtribuicaoProfessor[] — o que a turma recebeu, na ordem de atribuição\n\n**Identidade:**\n\nO token: a turma tem de ser da conta.\n\n**Regras de negócio no back (a tela não calcula):**\n\nconcluidoPor = alunos da turma (Alunos.ClassID, Ativo = TRUE) com sessão concluída no exercício; totalAlunos = esses alunos. A tela não conta.\n\n**Notas:**\n\nA visão do aluno é GET /aluno/salas/:id, sem contagem da turma.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AtribuicaoProfessor[] — o que a turma recebeu, na ordem de atribuição",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": [
                      {
                        "exercicioId": "ex-prof-1",
                        "titulo": "Acentuação em foco",
                        "dificuldade": "medio",
                        "atribuidoEm": "2026-02-03",
                        "concluidoPor": 3,
                        "totalAlunos": 4
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — não existe ou é de outra conta",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Turmas (professor)"
        ],
        "operationId": "turmas.atribuir",
        "summary": "atribuir: Atribuicao[] — a lista crua de atribuições da turma, como ficou",
        "description": "**Chamada no front:** `api.turmas.atribuir`\n\n**Corpo:** { exercicioIds } — vários de uma vez\n\n**Resposta:** 200 Atribuicao[] — a lista crua de atribuições da turma, como ficou\n\n**Identidade:**\n\nO token: a turma E cada exercício têm de ser da conta.\n\n**Regras de negócio no back (a tela não calcula):**\n\nConfere todos os ids antes de gravar qualquer um. Repetir um já atribuído não duplica: a chave de AtribuicoesProf é o par (ClassID, ExerciseID).\n\n**Notas:**\n\nA resposta não traz título nem concluidoPor: a tela relê GET /turmas/:id/atribuicoes depois.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "{ exercicioIds } — vários de uma vez",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "exemplo1": {
                  "summary": "Exemplo 1",
                  "value": {
                    "exercicioIds": [
                      "ex-prof-3",
                      "ex-prof-4"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atribuicao[] — a lista crua de atribuições da turma, como ficou",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": [
                      {
                        "exerciseId": "ex-prof-1",
                        "atribuidoEm": "2026-02-03",
                        "prazo": null
                      },
                      {
                        "exerciseId": "ex-prof-3",
                        "atribuidoEm": "2026-10-03",
                        "prazo": null
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "`DADOS_INVALIDOS` — exercicioIds vazio, que não é lista de ids, ou o corpo traz professorId",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "DADOS_INVALIDOS",
                    "value": {
                      "mensagem": "...",
                      "codigo": "DADOS_INVALIDOS"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma ou algum exercício não é da conta (ou está arquivado)",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/turmas/{id}/atribuicoes/{exercicioId}": {
      "delete": {
        "tags": [
          "Turmas (professor)"
        ],
        "operationId": "turmas.removerAtribuicao",
        "summary": "removerAtribuicao: (sem corpo)",
        "description": "**Chamada no front:** `api.turmas.removerAtribuicao`\n\n**Resposta:** 204 (sem corpo)\n\n**Identidade:**\n\nO token: a turma tem de ser da conta.\n\n**Notas:**\n\nA atribuição não tem id próprio: a remoção é pelo PAR turma + exercício, a chave primária de AtribuicoesProf.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exercicioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "(sem corpo)"
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma de outra conta, ou o exercício não está atribuído a ela",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/turmas/{id}/alunos/{alunoId}": {
      "delete": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.remover",
        "summary": "remover: (sem corpo)",
        "description": "**Chamada no front:** `api.alunos.remover`\n\n**Resposta:** 204 (sem corpo)\n\n**Identidade:**\n\nO token: a turma tem de ser da conta (senão 404). O :alunoId é o aluno-ALVO — sobre quem a ação é —, e não quem pede.\n\n**Regras de negócio no back (a tela não calcula):**\n\nApaga a linha de Alunos; as sessões dele vão junto (SessionsProf tem ON DELETE CASCADE). Ele deixa de conseguir entrar na hora.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "(sem corpo)"
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma de outra conta, ou o aluno não é desta turma",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}